+++ src/java/com/zimbra/cs/db/DbImapFolder.java
@@ -222,7 +222,7 @@ public class DbImapFolder {
     }
 
 
-    public static String getTableName(int mailboxId, int groupId) {
+    public static String getTableName(long mailboxId, long groupId) {
         return DbMailbox.qualifyTableName(groupId, TABLE_IMAP_FOLDER);
     }
