+++ src/java/com/zimbra/cs/redolog/op/ModifyContact.java
@@ -32,9 +32,6 @@ import com.zimbra.cs.mime.ParsedContact;
 import com.zimbra.cs.redolog.RedoLogInput;
 import com.zimbra.cs.redolog.RedoLogOutput;
 
-/**
- * @author jhahm
- */
 public class ModifyContact extends RedoableOp {
 
     private int mId;
@@ -51,7 +48,7 @@ public class ModifyContact extends RedoableOp {
         mId = UNKNOWN_ID;
     }
 
-    public ModifyContact(int mailboxId, int id, ParsedContact pc) {
+    public ModifyContact(long mailboxId, int id, ParsedContact pc) {
         setMailboxId(mailboxId);
         mId = id;
         mFields = pc.getFields();
